+++ CHANGES.md
@@ -1,5 +1,10 @@
 # CHANGES for stickynote
 
+## 0.2.0
+
+- feat: allow note (string record) metadata with optional filter argument
+    on getAllNotes and clear
+
 ## 0.1.1
 
 - fix: CJS export path